re-structure the project
